fix: Algolia rework cleanups (#234)
