+++ utils/assetChecks.ts
@@ -1,17 +1,12 @@
-import { nftStandards, tokenStandards } from '@/types/assets'
-
-import type { Asset } from '@/models/asset'
-
 export const isLyx = (asset?: Asset) => asset?.isNativeToken
 
 export const isNft = (asset?: Asset) =>
-  asset?.standard && nftStandards.includes(asset.standard)
+  asset?.standard && NFT_STANDARDS.includes(asset.standard)
 
 export const isToken = (asset?: Asset) =>
-  (asset?.standard && tokenStandards.includes(asset.standard)) ||
+  (asset?.standard && TOKEN_STANDARDS.includes(asset.standard)) ||
   asset?.isNativeToken
 
 export const isLsp7 = (asset?: Asset) => asset?.standard === 'LSP7DigitalAsset'
 
-export const isLsp8 = (asset?: Asset) =>
-  asset?.standard === 'LSP8IdentifiableDigitalAsset'
+export const isLsp8 = (asset?: Asset) => asset?.standard === 'LSP8DigitalAsset'
